refactor(server): update templates to import Nodes for triggers/tools/MCP; adjust paths (#211)
