+++ apps/server/src/nodes/mcp/local/local-mcp-server.node.ts
