+++ apps/server/src/templates.ts
@@ -2,27 +2,27 @@ import { toJSONSchema } from 'zod';
 import { Agent, AgentStaticConfigSchema } from './nodes/agent.node';
 import { ContainerProviderEntity, ContainerProviderExposedStaticConfigSchema } from './entities/containerProvider.entity';
 import { TemplateRegistry } from './graph';
-import { LocalMCPServer } from './mcp';
+import { LocalMCPServer } from './nodes/mcp/local/local-mcp-server.node';
 import { CheckpointerService } from './services/checkpointer.service';
 import { ConfigService } from './services/config.service';
 import { ContainerService } from './services/container.service';
 import { LoggerService } from './services/logger.service';
 import { VaultService, VaultConfigSchema } from './services/vault.service';
-import { CallAgentTool, CallAgentToolStaticConfigSchema } from './tools/call_agent.tool';
-import { GithubCloneRepoTool, GithubCloneRepoToolExposedStaticConfigSchema } from './tools/github_clone_repo';
-import { SendSlackMessageTool, SendSlackMessageToolExposedStaticConfigSchema } from './tools/send_slack_message.tool';
-import { ShellTool, ShellToolStaticConfigSchema } from './tools/shell_command';
-import { SlackTrigger } from './triggers';
-import { SlackTriggerExposedStaticConfigSchema } from './triggers/slack.trigger';
-import { RemindMeTool, RemindMeToolStaticConfigSchema } from './tools/remind_me.tool';
-import { DebugToolTrigger, DebugToolTriggerStaticConfigSchema } from './triggers/debugTool.trigger';
-import { LocalMcpServerStaticConfigSchema } from './mcp/localMcpServer';
-import { FinishTool, FinishToolStaticConfigSchema } from './tools/finish.tool';
+import { CallAgentTool, CallAgentToolStaticConfigSchema } from './nodes/tools/call-agent-tool.node';
+import { GithubCloneRepoTool, GithubCloneRepoToolExposedStaticConfigSchema } from './nodes/tools/github-clone-repo-tool.node';
+import { SendSlackMessageTool, SendSlackMessageToolExposedStaticConfigSchema } from './nodes/tools/send-slack-message-tool.node';
+import { ShellTool, ShellToolStaticConfigSchema } from './nodes/tools/shell-tool.node';
+import { SlackTrigger } from './nodes/triggers/slack.trigger.node';
+import { SlackTriggerExposedStaticConfigSchema } from './nodes/triggers/slack.trigger.node';
+import { RemindMeTool, RemindMeToolStaticConfigSchema } from './nodes/tools/remind-me-tool.node';
+import { DebugToolTrigger, DebugToolTriggerStaticConfigSchema } from './nodes/triggers/debug-tool.trigger.node';
+import { LocalMcpServerStaticConfigSchema } from './nodes/mcp/local/local-mcp-server.node';
+import { FinishTool, FinishToolStaticConfigSchema } from './nodes/tools/finish-tool.node';
 import { MongoService } from './services/mongo.service';
 import { MemoryNode, MemoryNodeStaticConfigSchema } from './nodes/memory.node';
 import { MemoryConnectorNode, MemoryConnectorStaticConfigSchema } from './nodes/memory.connector.node';
 // Unified Memory tool
-import { UnifiedMemoryTool, UnifiedMemoryToolNodeStaticConfigSchema } from './tools/memory/memory.tool';
+import { UnifiedMemoryTool, UnifiedMemoryToolNodeStaticConfigSchema } from './nodes/tools/memory-tool.node';
 
 export interface TemplateRegistryDeps {
   logger: LoggerService;
